[guppy] move fixtures into their own crate
As part of the test plan for guppy summaries, we're going to
write a small tool to generate a few summaries for each fixture.
