+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "cargo-guppy",
+    "fixtures",
     "guppy",
     "guppy-cmdlib",
     "guppy-summaries",
