+++ fixtures/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "fixtures"
+version = "0.1.0"
+authors = ["Rain <rain1@calibra.com>"]
+publish = false
+edition = "2018"
+
+[dependencies]
+guppy = { path = "../guppy" }
+once_cell = "1.4.0"
+pretty_assertions = "0.6.1"
