+++ fixtures/src/json.rs
@@ -1,159 +1,142 @@
 // Copyright (c) The cargo-guppy Contributors
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use crate::errors::FeatureBuildStage;
-use crate::graph::{
-    BuildTargetId, BuildTargetKind, DependencyDirection, EnabledStatus, EnabledTernary,
-    PackageGraph, PackageLink, PackageMetadata, PackageSource, Workspace,
-};
-use crate::unit_tests::dep_helpers::{
-    assert_all_links, assert_deps_internal, assert_topo_ids, assert_topo_metadatas,
-    assert_transitive_deps_internal,
-};
-use crate::{errors::FeatureGraphWarning, CargoMetadata, DependencyKind, PackageId, Platform};
+use crate::details::{FixtureDetails, LinkDetails, PackageDetails, PlatformResults};
+use crate::package_id;
+use guppy::errors::FeatureBuildStage;
+use guppy::graph::{BuildTargetId, BuildTargetKind, EnabledTernary, PackageGraph};
+use guppy::{errors::FeatureGraphWarning, CargoMetadata, DependencyKind, Platform, TargetFeatures};
 use once_cell::sync::Lazy;
-use pretty_assertions::assert_eq;
-use semver::Version;
-use std::collections::{BTreeMap, HashMap};
-use std::path::{Path, PathBuf};
-use target_spec::TargetFeatures;
+use std::collections::HashMap;
 
 // Metadata along with interesting crate names.
-pub(crate) static METADATA1: &str = include_str!("../../../fixtures/small/metadata1.json");
-pub(crate) static METADATA1_TESTCRATE: &str = "testcrate 0.1.0 (path+file:///fakepath/testcrate)";
-pub(crate) static METADATA1_DATATEST: &str =
+pub static METADATA1: &str = include_str!("../small/metadata1.json");
+pub static METADATA1_TESTCRATE: &str = "testcrate 0.1.0 (path+file:///fakepath/testcrate)";
+pub static METADATA1_DATATEST: &str =
     "datatest 0.4.2 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA1_REGION: &str =
+pub static METADATA1_REGION: &str =
     "region 2.1.2 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA1_DTOA: &str =
+pub static METADATA1_DTOA: &str =
     "dtoa 0.4.4 (registry+https://github.com/rust-lang/crates.io-index)";
 
-pub(crate) static METADATA2: &str = include_str!("../../../fixtures/small/metadata2.json");
-pub(crate) static METADATA2_TESTCRATE: &str =
+pub static METADATA2: &str = include_str!("../small/metadata2.json");
+pub static METADATA2_TESTCRATE: &str =
     "testworkspace-crate 0.1.0 (path+file:///Users/fakeuser/local/testworkspace/testcrate)";
-pub(crate) static METADATA2_WALKDIR: &str =
+pub static METADATA2_WALKDIR: &str =
     "walkdir 2.2.9 (path+file:///Users/fakeuser/local/testworkspace/walkdir)";
-pub(crate) static METADATA2_QUOTE: &str = "quote 1.0.2 (path+file:///Users/fakeuser/local/quote)";
+pub static METADATA2_QUOTE: &str = "quote 1.0.2 (path+file:///Users/fakeuser/local/quote)";
 
-pub(crate) static METADATA_DUPS: &str = include_str!("../../../fixtures/small/metadata_dups.json");
-pub(crate) static METADATA_DUPS_TESTCRATE: &str =
+pub static METADATA_DUPS: &str = include_str!("../small/metadata_dups.json");
+pub static METADATA_DUPS_TESTCRATE: &str =
     "testcrate-dups 0.1.0 (path+file:///Users/fakeuser/local/testcrates/testcrate-dups)";
-pub(crate) static METADATA_DUPS_LAZY_STATIC_1: &str =
+pub static METADATA_DUPS_LAZY_STATIC_1: &str =
     "lazy_static 1.4.0 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_DUPS_LAZY_STATIC_02: &str =
+pub static METADATA_DUPS_LAZY_STATIC_02: &str =
     "lazy_static 0.2.11 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_DUPS_BYTES_03: &str =
+pub static METADATA_DUPS_BYTES_03: &str =
     "bytes 0.3.0 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_DUPS_BYTES_05: &str =
+pub static METADATA_DUPS_BYTES_05: &str =
     "bytes 0.5.4 (registry+https://github.com/rust-lang/crates.io-index)";
 
-pub(crate) static METADATA_CYCLE1: &str =
-    include_str!("../../../fixtures/small/metadata_cycle1.json");
-pub(crate) static METADATA_CYCLE1_BASE: &str =
+pub static METADATA_CYCLE1: &str = include_str!("../small/metadata_cycle1.json");
+pub static METADATA_CYCLE1_BASE: &str =
     "testcycles-base 0.1.0 (path+file:///Users/fakeuser/local/testcrates/testcycles/testcycles-base)";
-pub(crate) static METADATA_CYCLE1_HELPER: &str =
+pub static METADATA_CYCLE1_HELPER: &str =
     "testcycles-helper 0.1.0 (path+file:///Users/fakeuser/local/testcrates/testcycles/testcycles-helper)";
 
-pub(crate) static METADATA_CYCLE2: &str =
-    include_str!("../../../fixtures/small/metadata_cycle2.json");
-pub(crate) static METADATA_CYCLE2_UPPER_A: &str =
+pub static METADATA_CYCLE2: &str = include_str!("../small/metadata_cycle2.json");
+pub static METADATA_CYCLE2_UPPER_A: &str =
     "upper-a 0.1.0 (path+file:///Users/fakeuser/local/testcrates/cycle2/upper-a)";
-pub(crate) static METADATA_CYCLE2_UPPER_B: &str =
+pub static METADATA_CYCLE2_UPPER_B: &str =
     "upper-b 0.1.0 (path+file:///Users/fakeuser/local/testcrates/cycle2/upper-b)";
-pub(crate) static METADATA_CYCLE2_LOWER_A: &str =
+pub static METADATA_CYCLE2_LOWER_A: &str =
     "lower-a 0.1.0 (path+file:///Users/fakeuser/local/testcrates/cycle2/lower-a)";
-pub(crate) static METADATA_CYCLE2_LOWER_B: &str =
+pub static METADATA_CYCLE2_LOWER_B: &str =
     "lower-b 0.1.0 (path+file:///Users/fakeuser/local/testcrates/cycle2/lower-b)";
 
-pub(crate) static METADATA_TARGETS1: &str =
-    include_str!("../../../fixtures/small/metadata_targets1.json");
-pub(crate) static METADATA_TARGETS1_TESTCRATE: &str =
+pub static METADATA_TARGETS1: &str = include_str!("../small/metadata_targets1.json");
+pub static METADATA_TARGETS1_TESTCRATE: &str =
     "testcrate-targets 0.1.0 (path+file:///Users/fakeuser/local/testcrates/testcrate-targets)";
-pub(crate) static METADATA_TARGETS1_LAZY_STATIC_1: &str =
+pub static METADATA_TARGETS1_LAZY_STATIC_1: &str =
     "lazy_static 1.4.0 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_TARGETS1_LAZY_STATIC_02: &str =
+pub static METADATA_TARGETS1_LAZY_STATIC_02: &str =
     "lazy_static 0.2.11 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_TARGETS1_LAZY_STATIC_01: &str =
+pub static METADATA_TARGETS1_LAZY_STATIC_01: &str =
     "lazy_static 0.1.16 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_TARGETS1_BYTES: &str =
+pub static METADATA_TARGETS1_BYTES: &str =
     "bytes 0.5.3 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_TARGETS1_DEP_A: &str =
+pub static METADATA_TARGETS1_DEP_A: &str =
     "dep-a 0.1.0 (path+file:///Users/fakeuser/local/testcrates/dep-a)";
 
-pub(crate) static METADATA_BUILD_TARGETS1: &str =
-    include_str!("../../../fixtures/small/metadata_build_targets1.json");
-pub(crate) static METADATA_BUILD_TARGETS1_TESTCRATE: &str =
+pub static METADATA_BUILD_TARGETS1: &str = include_str!("../small/metadata_build_targets1.json");
+pub static METADATA_BUILD_TARGETS1_TESTCRATE: &str =
     "testcrate 0.1.0 (path+file:///Users/fakeuser/local/testcrates/test-build-targets)";
 
-pub(crate) static METADATA_PROC_MACRO1: &str =
-    include_str!("../../../fixtures/small/metadata_proc_macro1.json");
-pub(crate) static METADATA_PROC_MACRO1_MACRO: &str =
+pub static METADATA_PROC_MACRO1: &str = include_str!("../small/metadata_proc_macro1.json");
+pub static METADATA_PROC_MACRO1_MACRO: &str =
     "macro 0.1.0 (path+file:///Users/fakeuser/local/testcrates/proc-macro/macro)";
-pub(crate) static METADATA_PROC_MACRO1_NORMAL_USER: &str =
+pub static METADATA_PROC_MACRO1_NORMAL_USER: &str =
     "normal-user 0.1.0 (path+file:///Users/fakeuser/local/testcrates/proc-macro/normal-user)";
-pub(crate) static METADATA_PROC_MACRO1_BUILD_USER: &str =
+pub static METADATA_PROC_MACRO1_BUILD_USER: &str =
     "build-user 0.1.0 (path+file:///Users/fakeuser/local/testcrates/proc-macro/build-user)";
-pub(crate) static METADATA_PROC_MACRO1_DEV_USER: &str =
+pub static METADATA_PROC_MACRO1_DEV_USER: &str =
     "dev-user 0.1.0 (path+file:///Users/fakeuser/local/testcrates/proc-macro/dev-user)";
 
-pub(crate) static METADATA_LIBRA: &str =
-    include_str!("../../../fixtures/large/metadata_libra.json");
-pub(crate) static METADATA_LIBRA_ADMISSION_CONTROL_SERVICE: &str =
+pub static METADATA_LIBRA: &str = include_str!("../large/metadata_libra.json");
+pub static METADATA_LIBRA_ADMISSION_CONTROL_SERVICE: &str =
     "admission-control-service 0.1.0 (path+file:///Users/fakeuser/local/libra/admission_control/admission-control-service)";
-pub(crate) static METADATA_LIBRA_COMPILER: &str =
+pub static METADATA_LIBRA_COMPILER: &str =
     "compiler 0.1.0 (path+file:///Users/fakeuser/local/libra/language/compiler)";
-pub(crate) static METADATA_LIBRA_E2E_TESTS: &str =
+pub static METADATA_LIBRA_E2E_TESTS: &str =
     "language-e2e-tests 0.1.0 (path+file:///Users/fakeuser/local/libra/language/e2e-tests)";
-pub(crate) static METADATA_LIBRA_EXECUTOR: &str =
+pub static METADATA_LIBRA_EXECUTOR: &str =
     "executor 0.1.0 (path+file:///Users/fakeuser/local/libra/execution/executor)";
-pub(crate) static METADATA_LIBRA_EXECUTOR_UTILS: &str =
+pub static METADATA_LIBRA_EXECUTOR_UTILS: &str =
     "executor-utils 0.1.0 (path+file:///Users/fakeuser/local/libra/execution/executor-utils)";
-pub(crate) static METADATA_LIBRA_COST_SYNTHESIS: &str =
+pub static METADATA_LIBRA_COST_SYNTHESIS: &str =
     "cost-synthesis 0.1.0 (path+file:///Users/fakeuser/local/libra/language/tools/cost-synthesis)";
-pub(crate) static METADATA_LIBRA_FUNCTIONAL_TESTS: &str =
+pub static METADATA_LIBRA_FUNCTIONAL_TESTS: &str =
     "functional_tests 0.1.0 (path+file:///Users/fakeuser/local/libra/language/functional_tests)";
-pub(crate) static METADATA_LIBRA_FUNCTIONAL_HYPHEN_TESTS: &str =
+pub static METADATA_LIBRA_FUNCTIONAL_HYPHEN_TESTS: &str =
     "functional-tests 0.1.0 (path+file:///Users/fakeuser/local/libra/language/functional-tests)";
-pub(crate) static METADATA_LIBRA_LIBRA_VM: &str =
+pub static METADATA_LIBRA_LIBRA_VM: &str =
     "libra-vm 0.1.0 (path+file:///Users/fakeuser/local/libra/language/libra-vm)";
-pub(crate) static METADATA_LIBRA_MOVE_LANG: &str =
+pub static METADATA_LIBRA_MOVE_LANG: &str =
     "move-lang 0.0.1 (path+file:///Users/fakeuser/local/libra/language/move-lang)";
-pub(crate) static METADATA_LIBRA_MOVE_LANG_STDLIB: &str =
+pub static METADATA_LIBRA_MOVE_LANG_STDLIB: &str =
     "move-lang-stdlib 0.1.0 (path+file:///Users/fakeuser/local/libra/language/move-lang/stdlib)";
-pub(crate) static METADATA_LIBRA_MOVE_VM_RUNTIME: &str =
+pub static METADATA_LIBRA_MOVE_VM_RUNTIME: &str =
     "move-vm-runtime 0.1.0 (path+file:///Users/fakeuser/local/libra/language/move-vm/runtime)";
-pub(crate) static METADATA_LIBRA_STDLIB: &str =
+pub static METADATA_LIBRA_STDLIB: &str =
     "stdlib 0.1.0 (path+file:///Users/fakeuser/local/libra/language/stdlib)";
-pub(crate) static METADATA_LIBRA_TEST_GENERATION: &str =
+pub static METADATA_LIBRA_TEST_GENERATION: &str =
     "test-generation 0.1.0 (path+file:///Users/fakeuser/local/libra/language/tools/test-generation)";
-pub(crate) static METADATA_LIBRA_TRANSACTION_BUILDER: &str =
+pub static METADATA_LIBRA_TRANSACTION_BUILDER: &str =
     "transaction-builder 0.1.0 (path+file:///Users/fakeuser/local/libra/language/transaction-builder)";
-pub(crate) static METADATA_LIBRA_VM_GENESIS: &str =
+pub static METADATA_LIBRA_VM_GENESIS: &str =
     "vm-genesis 0.1.0 (path+file:///Users/fakeuser/local/libra/language/tools/vm-genesis)";
-pub(crate) static METADATA_LIBRA_LANGUAGE_BENCHMARKS: &str =
+pub static METADATA_LIBRA_LANGUAGE_BENCHMARKS: &str =
     "language_benchmarks 0.1.0 (path+file:///Users/fakeuser/local/libra/language/benchmarks)";
-pub(crate) static METADATA_LIBRA_TREE_HEAP: &str =
+pub static METADATA_LIBRA_TREE_HEAP: &str =
     "tree_heap 0.1.0 (path+file:///Users/fakeuser/local/libra/language/stackless-bytecode/tree_heap)";
-pub(crate) static METADATA_LIBRA_LAZY_STATIC: &str =
+pub static METADATA_LIBRA_LAZY_STATIC: &str =
     "lazy_static 1.4.0 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_LIBRA_BACKTRACE: &str =
+pub static METADATA_LIBRA_BACKTRACE: &str =
     "backtrace 0.3.37 (registry+https://github.com/rust-lang/crates.io-index)";
-pub(crate) static METADATA_LIBRA_CFG_IF: &str =
+pub static METADATA_LIBRA_CFG_IF: &str =
     "cfg-if 0.1.9 (registry+https://github.com/rust-lang/crates.io-index)";
 
-pub(crate) static METADATA_LIBRA_F0091A4: &str =
-    include_str!("../../../fixtures/large/metadata_libra_f0091a4.json");
+pub static METADATA_LIBRA_F0091A4: &str = include_str!("../large/metadata_libra_f0091a4.json");
 
-pub(crate) static METADATA_LIBRA_9FFD93B: &str =
-    include_str!("../../../fixtures/large/metadata_libra_9ffd93b.json");
+pub static METADATA_LIBRA_9FFD93B: &str = include_str!("../large/metadata_libra_9ffd93b.json");
 
-pub(crate) static FAKE_AUTHOR: &str = "Fake Author <fakeauthor@example.com>";
+pub static FAKE_AUTHOR: &str = "Fake Author <fakeauthor@example.com>";
 
 macro_rules! define_fixture {
     ($name: ident, $json: ident) => {
-        pub(crate) fn $name() -> &'static Fixture {
-            static FIXTURE: Lazy<Fixture> = Lazy::new(|| Fixture {
-                graph: Fixture::parse_graph($json),
+        pub fn $name() -> &'static JsonFixture {
+            static FIXTURE: Lazy<JsonFixture> = Lazy::new(|| JsonFixture {
+                graph: JsonFixture::parse_graph($json),
                 details: FixtureDetails::$name(),
             });
             &*FIXTURE
@@ -161,61 +144,31 @@ macro_rules! define_fixture {
     };
 }
 
-pub(crate) struct Fixture {
+pub struct JsonFixture {
     graph: PackageGraph,
     details: FixtureDetails,
 }
 
-#[derive(Clone, Debug)]
-pub(crate) struct PlatformResults {
-    // Each pair stands for (required on, enabled on).
-    status: (EnabledTernary, EnabledTernary),
-    default_features: (EnabledTernary, EnabledTernary),
-    feature_statuses: HashMap<String, (EnabledTernary, EnabledTernary)>,
-}
-
-impl PlatformResults {
-    fn new(
-        status: (EnabledTernary, EnabledTernary),
-        default_features: (EnabledTernary, EnabledTernary),
-    ) -> Self {
-        Self {
-            status,
-            default_features,
-            feature_statuses: HashMap::new(),
-        }
-    }
-
-    fn with_feature_status(
-        mut self,
-        feature: &str,
-        status: (EnabledTernary, EnabledTernary),
-    ) -> Self {
-        self.feature_statuses.insert(feature.to_string(), status);
-        self
-    }
-}
-
-impl Fixture {
+impl JsonFixture {
     /// Returns the package graph for this fixture.
-    pub(crate) fn graph(&self) -> &PackageGraph {
+    pub fn graph(&self) -> &PackageGraph {
         &self.graph
     }
 
     /// Returns a mutable reference to the package graph for this fixture.
     #[allow(dead_code)]
-    pub(crate) fn graph_mut(&mut self) -> &mut PackageGraph {
+    pub fn graph_mut(&mut self) -> &mut PackageGraph {
         &mut self.graph
     }
 
     /// Returns the test details for this fixture.
     #[allow(dead_code)]
-    pub(crate) fn details(&self) -> &FixtureDetails {
+    pub fn details(&self) -> &FixtureDetails {
         &self.details
     }
 
     /// Verifies that the parsed metadata matches known details.
-    pub(crate) fn verify(&self) {
+    pub fn verify(&self) {
         self.graph
             .verify()
             .expect("graph verification should succeed");
@@ -298,303 +251,7 @@ impl Fixture {
     }
 }
 
-/// This captures metadata fields that are relevant for tests. They are meant to be written out
-/// lazily as tests are filled out -- feel free to add more details as necessary!
-pub(crate) struct FixtureDetails {
-    workspace_members: Option<BTreeMap<PathBuf, PackageId>>,
-    package_details: HashMap<PackageId, PackageDetails>,
-    link_details: HashMap<(PackageId, PackageId), LinkDetails>,
-    feature_graph_warnings: Vec<FeatureGraphWarning>,
-    cycles: Vec<Vec<PackageId>>,
-}
-
 impl FixtureDetails {
-    pub(crate) fn new(package_details: HashMap<PackageId, PackageDetails>) -> Self {
-        Self {
-            workspace_members: None,
-            package_details,
-            link_details: HashMap::new(),
-            feature_graph_warnings: vec![],
-            cycles: vec![],
-        }
-    }
-
-    pub(crate) fn with_workspace_members<'a>(
-        mut self,
-        workspace_members: impl IntoIterator<Item = (impl Into<PathBuf>, &'a str)>,
-    ) -> Self {
-        self.workspace_members = Some(
-            workspace_members
-                .into_iter()
-                .map(|(path, id)| (path.into(), package_id(id)))
-                .collect(),
-        );
-        self
-    }
-
-    pub(crate) fn with_link_details<'a>(
-        mut self,
-        link_details: HashMap<(PackageId, PackageId), LinkDetails>,
-    ) -> Self {
-        self.link_details = link_details;
-        self
-    }
-
-    pub(crate) fn with_feature_graph_warnings(
-        mut self,
-        mut warnings: Vec<FeatureGraphWarning>,
-    ) -> Self {
-        warnings.sort();
-        self.feature_graph_warnings = warnings;
-        self
-    }
-
-    pub(crate) fn with_cycles(mut self, cycles: Vec<Vec<&'static str>>) -> Self {
-        let mut cycles: Vec<_> = cycles
-            .into_iter()
-            .map(|cycle| {
-                let mut cycle: Vec<_> = cycle.into_iter().map(package_id).collect();
-                cycle.sort();
-                cycle
-            })
-            .collect();
-        cycles.sort();
-        self.cycles = cycles;
-        self
-    }
-
-    pub(crate) fn known_ids<'a>(&'a self) -> impl Iterator<Item = &'a PackageId> + 'a {
-        self.package_details.keys()
-    }
-
-    pub(crate) fn assert_workspace(&self, workspace: Workspace) {
-        if let Some(expected_members) = &self.workspace_members {
-            let members: Vec<_> = workspace
-                .members()
-                .map(|(path, metadata)| (path, metadata.id()))
-                .collect();
-            assert_eq!(
-                expected_members
-                    .iter()
-                    .map(|(path, id)| (path.as_path(), id))
-                    .collect::<Vec<_>>(),
-                members,
-                "workspace members should be correct"
-            );
-
-            assert_eq!(
-                workspace.members().len(),
-                workspace.members_by_name().len(),
-                "workspace.members() and members_by_name() return the same number of items"
-            );
-            for (name, metadata) in workspace.members_by_name() {
-                assert_eq!(
-                    name,
-                    metadata.name(),
-                    "members_by_name returns consistent results"
-                );
-            }
-        }
-    }
-
-    pub(crate) fn assert_topo(&self, graph: &PackageGraph) {
-        assert_topo_ids(graph, DependencyDirection::Forward, "topo sort");
-        assert_topo_ids(graph, DependencyDirection::Reverse, "reverse topo sort");
-        assert_topo_metadatas(graph, DependencyDirection::Forward, "topo sort (metadatas)");
-        assert_topo_metadatas(
-            graph,
-            DependencyDirection::Reverse,
-            "reverse topo sort (metadatas)",
-        );
-        assert_all_links(graph, DependencyDirection::Forward, "all links");
-        assert_all_links(graph, DependencyDirection::Reverse, "all links reversed");
-    }
-
-    pub(crate) fn assert_metadata(&self, id: &PackageId, metadata: PackageMetadata<'_>, msg: &str) {
-        let details = &self.package_details[id];
-        details.assert_metadata(metadata, msg);
-    }
-
-    // ---
-    // Build targets
-    // ---
-
-    pub(crate) fn has_build_targets(&self, id: &PackageId) -> bool {
-        let details = &self.package_details[id];
-        details.build_targets.is_some()
-    }
-
-    pub(crate) fn assert_build_targets(&self, metadata: PackageMetadata<'_>, msg: &str) {
-        let build_targets = self.package_details[metadata.id()]
-            .build_targets
-            .as_ref()
-            .unwrap();
-
-        let mut actual: Vec<_> = metadata
-            .build_targets()
-            .map(|build_target| {
-                // Strip off the manifest path from the beginning.
-                let path = build_target
-                    .path()
-                    .strip_prefix(
-                        metadata
-                            .manifest_path()
-                            .parent()
-                            .expect("manifest path is a file"),
-                    )
-                    .expect("build target path is inside source dir")
-                    .to_path_buf();
-
-                (build_target.id(), build_target.kind().clone(), path)
-            })
-            .collect();
-        actual.sort();
-
-        assert_eq!(build_targets, &actual, "{}: build targets match", msg,);
-    }
-
-    // ---
-    // Direct dependencies
-    // ---
-
-    /// Returns true if the deps for this package are available to test against.
-    pub(crate) fn has_deps(&self, id: &PackageId) -> bool {
-        let details = &self.package_details[id];
-        details.deps.is_some()
-    }
-
-    pub(crate) fn assert_deps(&self, graph: &PackageGraph, id: &PackageId, msg: &str) {
-        let details = &self.package_details[id];
-        assert_deps_internal(&graph, DependencyDirection::Forward, details, msg);
-    }
-
-    /// Returns true if the reverse deps for this package are available to test against.
-    pub(crate) fn has_reverse_deps(&self, id: &PackageId) -> bool {
-        let details = &self.package_details[id];
-        details.reverse_deps.is_some()
-    }
-
-    pub(crate) fn assert_reverse_deps(&self, graph: &PackageGraph, id: &PackageId, msg: &str) {
-        let details = &self.package_details[id];
-        assert_deps_internal(&graph, DependencyDirection::Reverse, details, msg);
-    }
-
-    // ---
-    // Transitive dependencies
-    // ---
-
-    /// Returns true if the transitive deps for this package are available to test against.
-    pub(crate) fn has_transitive_deps(&self, id: &PackageId) -> bool {
-        let details = &self.package_details[id];
-        details.transitive_deps.is_some()
-    }
-
-    pub(crate) fn assert_transitive_deps<'a>(
-        &self,
-        graph: &PackageGraph,
-        id: &PackageId,
-        msg: &str,
-    ) {
-        assert_transitive_deps_internal(
-            graph,
-            DependencyDirection::Forward,
-            &self.package_details[id],
-            msg,
-        )
-    }
-
-    /// Returns true if the transitive reverse deps for this package are available to test against.
-    pub(crate) fn has_transitive_reverse_deps(&self, id: &PackageId) -> bool {
-        let details = &self.package_details[id];
-        details.transitive_reverse_deps.is_some()
-    }
-
-    pub(crate) fn assert_transitive_reverse_deps<'a>(
-        &self,
-        graph: &PackageGraph,
-        id: &PackageId,
-        msg: &str,
-    ) {
-        assert_transitive_deps_internal(
-            graph,
-            DependencyDirection::Reverse,
-            &self.package_details[id],
-            msg,
-        )
-    }
-
-    // ---
-    // Links
-    // ---
-
-    pub(crate) fn assert_link_details(&self, graph: &PackageGraph, msg: &str) {
-        for ((from, to), details) in &self.link_details {
-            let metadata = graph
-                .metadata(from)
-                .unwrap_or_else(|| panic!("{}: known package ID '{}' should be valid", msg, from));
-            let mut links: Vec<_> = metadata
-                .direct_links()
-                .filter(|link| link.to().id() == to)
-                .collect();
-            assert_eq!(
-                links.len(),
-                1,
-                "{}: exactly 1 link between '{}' and '{}'",
-                msg,
-                from,
-                to
-            );
-
-            let link = links.pop().unwrap();
-            let msg = format!("{}: {} -> {}", msg, from, to);
-            details.assert_metadata(link, &msg);
-        }
-    }
-
-    // ---
-    // Features
-    // ---
-
-    pub(crate) fn has_named_features(&self, id: &PackageId) -> bool {
-        self.package_details[id].named_features.is_some()
-    }
-
-    pub(crate) fn assert_named_features(&self, graph: &PackageGraph, id: &PackageId, msg: &str) {
-        let mut actual: Vec<_> = graph
-            .metadata(id)
-            .expect("package id should be valid")
-            .named_features()
-            .collect();
-        actual.sort();
-        let expected = self.package_details[id].named_features.as_ref().unwrap();
-        assert_eq!(expected, &actual, "{}", msg);
-    }
-
-    pub(crate) fn assert_feature_graph_warnings(&self, graph: &PackageGraph, msg: &str) {
-        let mut actual: Vec<_> = graph.feature_graph().build_warnings().to_vec();
-        actual.sort();
-        assert_eq!(&self.feature_graph_warnings, &actual, "{}", msg);
-    }
-
-    // ---
-    // Cycles
-    // ---
-
-    pub(crate) fn assert_cycles(&self, graph: &PackageGraph, msg: &str) {
-        let mut actual: Vec<_> = graph
-            .cycles()
-            .all_cycles()
-            .map(|cycle| {
-                let mut cycle: Vec<_> = cycle.into_iter().collect();
-                cycle.sort();
-                cycle
-            })
-            .collect();
-        actual.sort();
-
-        assert_eq!(&self.cycles, &actual, "{}", msg);
-    }
-
     // Specific fixtures follow.
 
     pub(crate) fn metadata1() -> Self {
@@ -1484,282 +1141,4 @@ impl FixtureDetails {
             vec![METADATA_LIBRA_EXECUTOR, METADATA_LIBRA_EXECUTOR_UTILS],
         ])
     }
-}
-
-pub(crate) struct PackageDetails {
-    id: PackageId,
-    name: &'static str,
-    version: Version,
-    authors: Vec<&'static str>,
-    description: Option<&'static str>,
-    license: Option<&'static str>,
-
-    source: Option<PackageSource<'static>>,
-    build_targets: Option<Vec<(BuildTargetId<'static>, BuildTargetKind<'static>, PathBuf)>>,
-    // The vector items are (name, package id).
-    // XXX add more details about dependency edges here?
-    deps: Option<Vec<(&'static str, PackageId)>>,
-    reverse_deps: Option<Vec<(&'static str, PackageId)>>,
-    transitive_deps: Option<Vec<PackageId>>,
-    transitive_reverse_deps: Option<Vec<PackageId>>,
-    named_features: Option<Vec<&'static str>>,
-}
-
-impl PackageDetails {
-    fn new(
-        id: &'static str,
-        name: &'static str,
-        version: &'static str,
-        authors: Vec<&'static str>,
-        description: Option<&'static str>,
-        license: Option<&'static str>,
-    ) -> Self {
-        Self {
-            id: package_id(id),
-            name,
-            version: Version::parse(version).expect("version should be valid"),
-            authors,
-            description,
-            license,
-            source: None,
-            build_targets: None,
-            deps: None,
-            reverse_deps: None,
-            transitive_deps: None,
-            transitive_reverse_deps: None,
-            named_features: None,
-        }
-    }
-
-    fn with_workspace_path(mut self, path: &'static str) -> Self {
-        self.source = Some(PackageSource::Workspace(Path::new(path)));
-        self
-    }
-
-    fn with_local_path(mut self, path: &'static str) -> Self {
-        self.source = Some(PackageSource::Path(Path::new(path)));
-        self
-    }
-
-    fn with_crates_io(self) -> Self {
-        self.with_external_source(PackageSource::CRATES_IO_REGISTRY)
-    }
-
-    fn with_external_source(mut self, source: &'static str) -> Self {
-        self.source = Some(PackageSource::External(source));
-        self
-    }
-
-    fn with_build_targets(
-        mut self,
-        mut build_targets: Vec<(
-            BuildTargetId<'static>,
-            BuildTargetKind<'static>,
-            &'static str,
-        )>,
-    ) -> Self {
-        build_targets.sort();
-        self.build_targets = Some(
-            build_targets
-                .into_iter()
-                .map(|(id, kind, path)| (id, kind, path.to_string().into()))
-                .collect(),
-        );
-        self
-    }
-
-    fn with_deps(mut self, mut deps: Vec<(&'static str, &'static str)>) -> Self {
-        deps.sort();
-        self.deps = Some(
-            deps.into_iter()
-                .map(|(name, id)| (name, package_id(id)))
-                .collect(),
-        );
-        self
-    }
-
-    fn with_reverse_deps(mut self, mut reverse_deps: Vec<(&'static str, &'static str)>) -> Self {
-        reverse_deps.sort();
-        self.reverse_deps = Some(
-            reverse_deps
-                .into_iter()
-                .map(|(name, id)| (name, package_id(id)))
-                .collect(),
-        );
-        self
-    }
-
-    fn with_transitive_deps(mut self, mut transitive_deps: Vec<&'static str>) -> Self {
-        transitive_deps.sort();
-        self.transitive_deps = Some(transitive_deps.into_iter().map(package_id).collect());
-        self
-    }
-
-    fn with_transitive_reverse_deps(
-        mut self,
-        mut transitive_reverse_deps: Vec<&'static str>,
-    ) -> Self {
-        transitive_reverse_deps.sort();
-        self.transitive_reverse_deps = Some(
-            transitive_reverse_deps
-                .into_iter()
-                .map(package_id)
-                .collect(),
-        );
-        self
-    }
-
-    fn with_named_features(mut self, mut named_features: Vec<&'static str>) -> Self {
-        named_features.sort();
-        self.named_features = Some(named_features);
-        self
-    }
-
-    fn insert_into(self, map: &mut HashMap<PackageId, PackageDetails>) {
-        map.insert(self.id.clone(), self);
-    }
-
-    pub(crate) fn id(&self) -> &PackageId {
-        &self.id
-    }
-
-    pub(crate) fn deps(
-        &self,
-        direction: DependencyDirection,
-    ) -> Option<&[(&'static str, PackageId)]> {
-        match direction {
-            DependencyDirection::Forward => self.deps.as_ref().map(|deps| deps.as_slice()),
-            DependencyDirection::Reverse => self.reverse_deps.as_ref().map(|deps| deps.as_slice()),
-        }
-    }
-
-    pub(crate) fn transitive_deps(&self, direction: DependencyDirection) -> Option<&[PackageId]> {
-        match direction {
-            DependencyDirection::Forward => {
-                self.transitive_deps.as_ref().map(|deps| deps.as_slice())
-            }
-            DependencyDirection::Reverse => self
-                .transitive_reverse_deps
-                .as_ref()
-                .map(|deps| deps.as_slice()),
-        }
-    }
-
-    pub(crate) fn assert_metadata(&self, metadata: PackageMetadata<'_>, msg: &str) {
-        assert_eq!(&self.id, metadata.id(), "{}: same package ID", msg);
-        assert_eq!(self.name, metadata.name(), "{}: same name", msg);
-        assert_eq!(&self.version, metadata.version(), "{}: same version", msg);
-        assert_eq!(
-            &self.authors,
-            &metadata
-                .authors()
-                .iter()
-                .map(|author| author.as_str())
-                .collect::<Vec<_>>(),
-            "{}: same authors",
-            msg
-        );
-        assert_eq!(
-            &self.description,
-            &metadata.description(),
-            "{}: same description",
-            msg
-        );
-        assert_eq!(&self.license, &metadata.license(), "{}: same license", msg);
-        if let Some(source) = &self.source {
-            assert_eq!(source, &metadata.source(), "{}: same source", msg);
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct LinkDetails {
-    from: PackageId,
-    to: PackageId,
-    platform_results: Vec<(DependencyKind, Platform<'static>, PlatformResults)>,
-    features: Vec<(DependencyKind, Vec<&'static str>)>,
-}
-
-impl LinkDetails {
-    pub(crate) fn new(from: PackageId, to: PackageId) -> Self {
-        Self {
-            from,
-            to,
-            platform_results: vec![],
-            features: vec![],
-        }
-    }
-
-    pub(crate) fn with_platform_status(
-        mut self,
-        dep_kind: DependencyKind,
-        platform: Platform<'static>,
-        status: PlatformResults,
-    ) -> Self {
-        self.platform_results.push((dep_kind, platform, status));
-        self
-    }
-
-    pub(crate) fn with_features(
-        mut self,
-        dep_kind: DependencyKind,
-        mut features: Vec<&'static str>,
-    ) -> Self {
-        features.sort();
-        self.features.push((dep_kind, features));
-        self
-    }
-
-    pub(crate) fn insert_into(self, map: &mut HashMap<(PackageId, PackageId), Self>) {
-        map.insert((self.from.clone(), self.to.clone()), self);
-    }
-
-    pub(crate) fn assert_metadata(&self, link: PackageLink<'_>, msg: &str) {
-        let required_enabled = |status: EnabledStatus<'_>, platform: &Platform<'_>| {
-            (status.required_on(platform), status.enabled_on(platform))
-        };
-
-        for (dep_kind, platform, results) in &self.platform_results {
-            let req = link.req_for_kind(*dep_kind);
-            assert_eq!(
-                required_enabled(req.status(), platform),
-                results.status,
-                "{}: for platform '{}', kind {}, status is correct",
-                msg,
-                platform.triple(),
-                dep_kind,
-            );
-            assert_eq!(
-                required_enabled(req.default_features(), platform),
-                results.default_features,
-                "{}: for platform '{}', kind {}, default features is correct",
-                msg,
-                platform.triple(),
-                dep_kind,
-            );
-            for (feature, status) in &results.feature_statuses {
-                assert_eq!(
-                    required_enabled(req.feature_status(feature), platform),
-                    *status,
-                    "{}: for platform '{}', kind {}, feature '{}' has correct status",
-                    msg,
-                    platform.triple(),
-                    dep_kind,
-                    feature
-                );
-            }
-        }
-
-        for (dep_kind, features) in &self.features {
-            let metadata = link.req_for_kind(*dep_kind);
-            let mut actual_features: Vec<_> = metadata.features().collect();
-            actual_features.sort();
-            assert_eq!(&actual_features, features, "{}: features is correct", msg);
-        }
-    }
-}
-
-/// Helper for creating `PackageId` instances in test code.
-pub(crate) fn package_id(s: impl Into<Box<str>>) -> PackageId {
-    PackageId::new(s)
 }
