+++ fixtures/src/lib.rs
@@ -1,0 +1,17 @@
+// Copyright (c) The cargo-guppy Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! Test fixtures for guppy.
+
+// TODO: move repo fixture code into this crate
+
+pub mod dep_helpers;
+pub mod details;
+pub mod json;
+
+use guppy::PackageId;
+
+/// Helper for creating `PackageId` instances in test code.
+pub fn package_id(s: impl Into<Box<str>>) -> PackageId {
+    PackageId::new(s)
+}
