+++ guppy/Cargo.toml
@@ -40,6 +40,7 @@ target-spec = { version = "0.2.4", path = "../target-spec" }
 
 [dev-dependencies]
 assert_matches = "1.3.0"
+fixtures = { path = "../fixtures" }
 pretty_assertions = "0.6.1"
 
 [features]
