+++ guppy/src/graph/feature/resolve.rs
@@ -381,8 +381,8 @@ impl<'g> FeatureSet<'g> {
     }
 
     // Currently a helper for debugging -- will be made public in the future.
-    #[allow(dead_code)]
-    pub(crate) fn links<'a>(
+    #[doc(hidden)]
+    pub fn links<'a>(
         &'a self,
         direction: DependencyDirection,
     ) -> impl Iterator<Item = (FeatureId<'g>, FeatureId<'g>, &'g FeatureEdge)> + 'a {
