+++ guppy/src/unit_tests/mod.rs
@@ -1,20 +1,4 @@
 // Copyright (c) The cargo-guppy Contributors
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-#[cfg(feature = "proptest010")]
-#[macro_use]
-mod proptest_helpers;
-
-#[cfg(not(feature = "proptest010"))]
-macro_rules! proptest_suite {
-    ($name: ident) => {
-        // Empty macro to skip proptests if the proptest feature is disabled.
-    };
-}
-
-mod dep_helpers;
 mod dot_tests;
-mod feature_helpers;
-mod fixtures;
-mod graph_tests;
-mod invalid_tests;
