+++ guppy/tests/graph-tests/feature_helpers.rs
@@ -1,8 +1,8 @@
 // Copyright (c) The cargo-guppy Contributors
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use crate::graph::feature::{FeatureList, FeatureSet};
-use crate::PackageId;
+use guppy::graph::feature::{FeatureList, FeatureSet};
+use guppy::PackageId;
 
 pub(super) fn assert_features_for_package(
     feature_set: &FeatureSet<'_>,
