+++ guppy/tests/graph-tests/main.rs
@@ -1,0 +1,17 @@
+// Copyright (c) The cargo-guppy Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+#[cfg(feature = "proptest010")]
+#[macro_use]
+mod proptest_helpers;
+
+#[cfg(not(feature = "proptest010"))]
+macro_rules! proptest_suite {
+    ($name: ident) => {
+        // Empty macro to skip proptests if the proptest feature is disabled.
+    };
+}
+
+mod feature_helpers;
+mod graph_tests;
+mod invalid_tests;
